Methods to Process and Wipe Devices

+++ src/Infrastructure/Writers/TransporterGroupWriter.cs
@@ -4,6 +4,11 @@
 public sealed class TransporterGroupWriter(IApplicationDbContext context) : ITransporterGroupWriter
 {
     // Create a new transporter group asynchronously
+    // Parameters:
+    // - transporterGroupDto: The transporter group data transfer object
+    // - cancellationToken: The cancellation token
+    // Returns:
+    // - The created transporter group view model
     public async Task<TransporterGroupVm> CreateTransporterGroupAsync(TransporterGroupDto transporterGroupDto, CancellationToken cancellationToken)
     {
         var transporterGroup = new TransporterGroup
@@ -21,14 +26,39 @@ public sealed class TransporterGroupWriter(IApplicationDbContext context) : ITra
     }
 
     // Delete a transporter group asynchronously
+    // Parameters:
+    // - transporterId: The ID of the transporter
+    // - groupId: The ID of the group
+    // - cancellationToken: The cancellation token
     public async Task DeleteTransporterGroupAsync(Guid transporterId, long groupId, CancellationToken cancellationToken)
     {
-        var transporterGroup = await context.TransportersGroup.FindAsync([transporterId, groupId], cancellationToken)
-            ?? throw new NotFoundException(nameof(TransporterGroup), $"{transporterId},{groupId}");
+        var transporterGroup = await context.TransportersGroup.FindAsync([transporterId, groupId], cancellationToken);
 
-        context.TransportersGroup.Attach(transporterGroup);
+        if (transporterGroup != default)
+        {
+            context.TransportersGroup.Attach(transporterGroup);
 
-        context.TransportersGroup.Remove(transporterGroup);
-        await context.SaveChangesAsync(cancellationToken);
+            context.TransportersGroup.Remove(transporterGroup);
+            await context.SaveChangesAsync(cancellationToken);
+        }
+    }
+
+    // Delete all transporter groups that match the given transporterId asynchronously
+    // Parameters:
+    // - transporterId: The ID of the transporter
+    // - cancellationToken: The cancellation token
+    public async Task DeleteTransporterGroupsAsync(Guid transporterId, CancellationToken cancellationToken)
+    {
+        var transporterGroups = await context.TransportersGroup
+            .Where(tg => tg.TransporterId == transporterId)
+            .ToListAsync(cancellationToken);
+
+        if (transporterGroups.Count != 0)
+        {
+            context.TransportersGroup.AttachRange(transporterGroups);
+
+            context.TransportersGroup.RemoveRange(transporterGroups);
+            await context.SaveChangesAsync(cancellationToken);
+        }
     }
 }
